+++ src/utils/messages/replyWarning.ts
@@ -43,6 +43,10 @@ export async function replyWarning(
 
 	if ((interaction.deferred || interaction.replied) && !components) {
 		if (!ephemeral) {
+			if (interaction.deferred) {
+				await interaction.editReply(messageToSend).catch((e) => null);
+				return;
+			}
 			if ("_reply" in interaction) {
 				await (await interaction.fetchReply()).delete().catch((e) => null);
 			} else {
